+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+
+## 46.5.2 - 2020-06-09
 - [Feature] Add new **Tile** component for displaying details and actions for various items ([#1350](https://github.com/optimizely/oui/pull/1350))
 - [Patch] Add `buttonRef` prop to **ButtonIcon** to be used as a dropdown activator ([#1350](https://github.com/optimizely/oui/pull/1350))
 - [Patch] Add `hasSpellCheck` prop to **Input** component ([#1338](https://github.com/optimizely/oui/pull/1338))